backend/requirements.txt
fastapi
uvicorn
pyodbc
pydantic

pymongo
mysql-connector-python
psycopg2-binary
sqlalchemy
